refactor (divide screens in individual packages)
